Putting source files into separate directory
